alx_travel_app/requirement.txt
asgiref==3.11.0
Django==5.2.8
django-cors-headers==4.9.0
django-seed==0.3.1
djangorestframework==3.16.1
drf-yasg==1.21.11
Faker==38.2.0
inflection==0.5.1
packaging==25.0
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.4
toposort==1.10
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
